♻️  refactor: [#30] util 함수로 분리
